+++ external/dendry/include/math2d.h
@@ -6,7 +6,7 @@
 
 #include "utils.h"
 
-const double EPS = 1e-9;
+const float EPS = 1e-9;
 
 struct Point2D;
 struct Vec2D;
@@ -17,12 +17,12 @@ struct Segment2D;
  */
 struct Point2D
 {
-	double x;
-	double y;
+	float x;
+	float y;
 
-	Point2D() : x(0.0), y(0.0) { }
+	Point2D() : x(0.f), y(0.f) { }
 
-	Point2D(double _x, double _y) : x(_x), y(_y) { }
+	Point2D(float _x, float _y) : x(_x), y(_y) { }
 
 	// Unary Point operators
 	Point2D& operator+=(const Point2D& p) { x += p.x; y += p.y; return *this; }
@@ -33,8 +33,8 @@ struct Point2D
 	Point2D& operator-=(const Vec2D& v);
 
 	// Scalar operators
-	Point2D& operator*=(double s) { x *= s; y *= s; return *this; }
-	Point2D& operator/=(double s) { x /= s; y /= s; return *this; }
+	Point2D& operator*=(float s) { x *= s; y *= s; return *this; }
+	Point2D& operator/=(float s) { x /= s; y /= s; return *this; }
 	
 	// Unary minus operator
 	Point2D operator-() const { return Point2D(-x, -y);	}
@@ -68,52 +68,52 @@ inline Point2D operator-(const Point2D& a, const Vec2D& v) {
 }
 
 // Binary scalar operators
-inline Point2D operator*(const Point2D& a, double s) {
+inline Point2D operator*(const Point2D& a, float s) {
 	return Point2D(a) *= s;
 }
 
-inline Point2D operator*(double s, const Point2D& a) {
+inline Point2D operator*(float s, const Point2D& a) {
 	return Point2D(a) *= s;
 }
 
-inline Point2D operator/(const Point2D& a, double s) {
+inline Point2D operator/(const Point2D& a, float s) {
 	return Point2D(a) /= s;
 }
 
 // Utility functions
-inline double dist_sq(const Point2D& lhs, const Point2D& rhs) {
+inline float dist_sq(const Point2D& lhs, const Point2D& rhs) {
 	return (lhs.x - rhs.x) * (lhs.x - rhs.x)
 		 + (lhs.y - rhs.y) * (lhs.y - rhs.y);
 }
 
-inline double dist(const Point2D& lhs, const Point2D& rhs) {
+inline float dist(const Point2D& lhs, const Point2D& rhs) {
 	return sqrt(dist_sq(lhs, rhs));
 }
 
-inline double hypot(const Point2D& lhs, const Point2D& rhs) {
+inline float hypot(const Point2D& lhs, const Point2D& rhs) {
 	return hypot(lhs.x - rhs.x, lhs.y - rhs.y);
 }
 
-inline Point2D lerp(const Point2D& a, const Point2D& b, double t) {
+inline Point2D lerp(const Point2D& a, const Point2D& b, float t) {
 	return Point2D(
 		lerp(a.x, b.x, t),
 		lerp(a.y, b.y, t)
 	);
 }
 
-double angle(const Point2D& a, const Point2D& o, const Point2D& b);
+float angle(const Point2D& a, const Point2D& o, const Point2D& b);
 
 /*
  * Vector in a 2D Space
  */
 struct Vec2D
 {
-	double x;
-	double y;
+	float x;
+	float y;
 
-	Vec2D() : x(0.0), y(0.0) { }
+	Vec2D() : x(0.f), y(0.f) { }
 
-	Vec2D(double _x, double _y) : x(_x), y(_y) { }
+	Vec2D(float _x, float _y) : x(_x), y(_y) { }
 
 	Vec2D(const Point2D& p) : x(p.x), y(p.y) { }
 
@@ -124,8 +124,8 @@ struct Vec2D
 	Vec2D& operator-=(const Vec2D& v) { x -= v.x; y -= v.y; return *this; }
 
 	// Scalar operators
-	Vec2D& operator*=(double s) { x *= s; y *= s; return *this; }
-	Vec2D& operator/=(double s) { x /= s; y /= s; return *this; }
+	Vec2D& operator*=(float s) { x *= s; y *= s; return *this; }
+	Vec2D& operator/=(float s) { x /= s; y /= s; return *this; }
 
 	// Unary minus operator
 	Vec2D operator-() const { return Vec2D(-x, -y); }
@@ -150,41 +150,41 @@ inline Vec2D operator-(const Vec2D& a, const Vec2D& b) {
 }
 
 // Binary scalar operators
-inline Vec2D operator*(const Vec2D& a, double s) {
+inline Vec2D operator*(const Vec2D& a, float s) {
 	return Vec2D(a) *= s;
 }
 
-inline Vec2D operator*(double s, const Vec2D& a) {
+inline Vec2D operator*(float s, const Vec2D& a) {
 	return Vec2D(a) *= s;
 }
 
-inline Vec2D operator/(const Vec2D& a, double s) {
+inline Vec2D operator/(const Vec2D& a, float s) {
 	return Vec2D(a) /= s;
 }
 
 // Utility functions
-inline double norm_sq(const Vec2D& a) {
+inline float norm_sq(const Vec2D& a) {
 	return a.x * a.x + a.y * a.y;
 }
 
-inline double norm(const Vec2D& a) {
+inline float norm(const Vec2D& a) {
 	return sqrt(norm_sq(a));
 }
 
-inline double hypot(const Vec2D& a) {
+inline float hypot(const Vec2D& a) {
 	return hypot(a.x, a.y);
 }
 
-inline double dot(const Vec2D& a, const Vec2D& b) {
+inline float dot(const Vec2D& a, const Vec2D& b) {
 	return a.x * b.x + a.y * b.y;
 }
 
-inline double cross(const Vec2D& a, const Vec2D& b) {
+inline float cross(const Vec2D& a, const Vec2D& b) {
 	return a.x * b.y - a.y * b.x;
 }
 
 inline Vec2D normalized(const Vec2D& a) {
-	const double n = norm(a);
+	const float n = norm(a);
 	return Vec2D(a.x / n, a.y / n);
 }
 
@@ -196,7 +196,7 @@ inline Vec2D rotateCW90(const Vec2D& v) {
 	return Vec2D(v.y, -v.x);
 }
 
-inline double angle(const Vec2D& oa, const Vec2D& ob) {
+inline float angle(const Vec2D& oa, const Vec2D& ob) {
 	return acos(dot(oa, ob) / sqrt(norm_sq(oa) * norm_sq(ob)));
 }
 
@@ -214,15 +214,15 @@ struct Segment2D
 };
 
 // Utility functions
-inline double length_sq(const Segment2D& s) {
+inline float length_sq(const Segment2D& s) {
 	return dist_sq(s.a, s.b);
 }
 
-inline double length(const Segment2D& s) {
+inline float length(const Segment2D& s) {
 	return dist(s.a, s.b);
 }
 
-inline Point2D lerp(const Segment2D& s, double t) {
+inline Point2D lerp(const Segment2D& s, float t) {
 	return lerp(s.a, s.b, t);
 }
 
@@ -240,22 +240,22 @@ std::array<Point2D, N> SubdivideInPoints(const Segment2D& s)
 
 	for (int n = 0; n < points.size(); n++)
 	{
-		const double t = double(n + 1) / (N + 1);
+		const float t = float(n + 1) / (N + 1);
 		points[n] = lerp(s.a, s.b, t);
 	}
 
 	return points;
 }
 
-double pointLineProjection(const Point2D& p, const Point2D& a, const Point2D& b);
-double pointLineProjection(const Point2D& p, const Segment2D& s);
+float pointLineProjection(const Point2D& p, const Point2D& a, const Point2D& b);
+float pointLineProjection(const Point2D& p, const Segment2D& s);
 
-double pointLineSegmentProjection(const Point2D& p, const Point2D& a, const Point2D& b);
-double pointLineSegmentProjection(const Point2D& p, const Segment2D& s);
+float pointLineSegmentProjection(const Point2D& p, const Point2D& a, const Point2D& b);
+float pointLineSegmentProjection(const Point2D& p, const Segment2D& s);
 
-double distToLine(const Point2D& p, const Point2D& a, const Point2D& b, Point2D& c);
+float distToLine(const Point2D& p, const Point2D& a, const Point2D& b, Point2D& c);
 
-double distToLineSegment(const Point2D& p, const Point2D& a, const Point2D& b, Point2D& c);
-double distToLineSegment(const Point2D& p, const Segment2D& s, Point2D& c);
+float distToLineSegment(const Point2D& p, const Point2D& a, const Point2D& b, Point2D& c);
+float distToLineSegment(const Point2D& p, const Segment2D& s, Point2D& c);
 
 #endif // MATH2D_H
