+++ external/dendry/source/math2d.cpp
@@ -16,23 +16,23 @@ Point2D& Point2D::operator-=(const Vec2D& v)
 	return *this;
 }
 
-double angle(const Point2D& a, const Point2D& o, const Point2D& b)
+float angle(const Point2D& a, const Point2D& o, const Point2D& b)
 {
 	const Vec2D oa(o, a);
 	const Vec2D ob(o, b);
 	return angle(oa, ob);
 }
 
-double pointLineProjection(const Point2D& p, const Point2D& a, const Point2D& b)
+float pointLineProjection(const Point2D& p, const Point2D& a, const Point2D& b)
 {
 	const Vec2D ap(a, p);
 	const Vec2D ab(a, b);
 
 	// Segment is only a point and has no length
-	if (norm_sq(ab) <= 0.0)
+	if (norm_sq(ab) <= 0.f)
 	{
 		// The nearest point on the segment is A (or B)
-		return 0.0;
+		return 0.f;
 	}
 
 	// Segment has a length greater than 0
@@ -40,45 +40,45 @@ double pointLineProjection(const Point2D& p, const Point2D& a, const Point2D& b)
 	return dot(ap, ab) / norm_sq(ab);
 }
 
-double pointLineProjection(const Point2D& p, const Segment2D& s)
+float pointLineProjection(const Point2D& p, const Segment2D& s)
 {
 	return pointLineProjection(p, s.a, s.b);
 }
 
-double pointLineSegmentProjection(const Point2D& p, const Point2D& a, const Point2D& b)
+float pointLineSegmentProjection(const Point2D& p, const Point2D& a, const Point2D& b)
 {
-	const double u = pointLineProjection(p, a, b);
-	return std::clamp(u, 0.0, 1.0);
+	const float u = pointLineProjection(p, a, b);
+	return std::clamp(u, 0.f, 1.f);
 }
 
-double pointLineSegmentProjection(const Point2D& p, const Segment2D& s)
+float pointLineSegmentProjection(const Point2D& p, const Segment2D& s)
 {
 	return pointLineSegmentProjection(p, s.a, s.b);
 }
 
-double distToLine(const Point2D& p, const Point2D& a, const Point2D& b, Point2D& c)
+float distToLine(const Point2D& p, const Point2D& a, const Point2D& b, Point2D& c)
 {
 	const Vec2D ab(a, b);
-	const double u = pointLineProjection(p, a, b);
+	const float u = pointLineProjection(p, a, b);
 
 	c = a + ab * u;
 
 	return dist(p, c);
 }
 
-double distToLineSegment(const Point2D& p, const Point2D& a, const Point2D& b, Point2D& c)
+float distToLineSegment(const Point2D& p, const Point2D& a, const Point2D& b, Point2D& c)
 {
 	const Vec2D ab(a, b);
-	const double u = pointLineProjection(p, a, b);
+	const float u = pointLineProjection(p, a, b);
 
-	if (u < 0.0)
+	if (u < 0.f)
 	{
 		// P is closer to A
 		c = a;
 		return dist(p, a);
 	}
 	
-	if (u > 1.0)
+	if (u > 1.f)
 	{
 		// P is closer to B
 		c = b;
@@ -90,7 +90,7 @@ double distToLineSegment(const Point2D& p, const Point2D& a, const Point2D& b, P
 	return dist(p, c);
 }
 
-double distToLineSegment(const Point2D& p, const Segment2D& s, Point2D& c)
+float distToLineSegment(const Point2D& p, const Segment2D& s, Point2D& c)
 {
 	return distToLineSegment(p, s.a, s.b, c);
 }
